Added lib/sojourner/utils/play.py, which consists of intermediate playbooks created

+++ lib/sojourner/utils/cli.py
@@ -6,16 +6,11 @@ from __future__ import (absolute_import, division, print_function)
 
 __metaclass__ = type
 
-import re,argparse,os,sys
-#import commands #Not using commands
-from subprocess import Popen, PIPE
+# +----------------------------------------------------------------------+
 
 from sojourner.utils.dbcon import *
+from sojourner.utils.play import *
 
-try:
-    import json
-except ImportError:
-    import simplejson as json
 # +----------------------------------------------------------------------+
 # This is for ansible inventory
 def grouplist():
@@ -102,11 +97,21 @@ def show (args) :
 # +----------------------------------------------------------------------+
 def reap (args) :
         machine=args.machine
-        query='DELETE FROM inventory'
-        query1 = query + " where Hostname='" + machine + "'"
-        data=execute_sql(query1)[1]
-        if (data==1):
-                print ("Successfully Removed %s from Inventory" %(machine))
+	debug=args.debug
+	fact_dest = "/etc/ansible/facts.d/sojourner.fact"
+	playbook=reap_local_fact(fact_dest)
+	hostfile=create_temp_inventory_file(machine)
+	status=run_playbook (hostfile,playbook,debug)
+	if status == 0:
+                #Cool playbook run successfull"
+        	query='DELETE FROM inventory'
+        	query1 = query + " where Hostname='" + machine + "'"
+        	data=execute_sql(query1)[1]
+        	if (data==1):
+                	print ("Successfully Removed %s from Inventory" %(machine))
+        else:
+                print ("Error:")
+                print (output)
         return 0
 # +----------------------------------------------------------------------+
 def assign(args):
@@ -127,97 +132,15 @@ def assign(args):
 	#Now we have confirmed that the role/cookbook path exists, lets set the environment
 	os.environ['ANSIBLE_ROLES_PATH'] = str(C.SOJOURNER_ANSIBLE_ROLES)
 	
-        # Creat a Temp inventory file for this server
-        content="[all]\n"
-        tab="\t"
-        content=content+machine+ "\tProduct=" + product + "\tRole=" + role + "\n"
-        hostfile="/tmp/"+machine+".yml"
-        fo = open(hostfile, "wb")
-        fo.write(content);
-        fo.close()
+	# Creat a Temp inventory file for this server
+	hostfile=create_temp_inventory_file(machine,product,role)	
 
-        playbook="/tmp/playbook/"
-        if not os.path.exists(playbook):
-                os.makedirs(playbook)
-        playbook=playbook+role+".yml"
+	fact_dest = "/etc/ansible/facts.d/sojourner.fact"
+	playbook=deploy_local_fact (role,fact_dest)
 	
-	dest = "/etc/ansible/facts.d/sojourner.fact"
-        content="""---
-- name: Deploy %s
-  hosts: all
-  user: root
-
-  vars :
-   - Dest : %s
-"""%(role,dest)
-
-        if C.SOJOURNER_PROVISIONER == 'chef':
-                content=content+"""
-  roles:
-    -  { role: chef_zero,cookbook: %s }
-""" %(role)
-        elif C.SOJOURNER_PROVISIONER == 'ansible':
-                content=content+"""
-  roles:
-    -  %s
-""" %(role)
-
-#Now the role section is added, lets add post_task
+	status=run_playbook (hostfile,playbook,debug)
 	
-	product_regex=r"'Product=((?!{{Product}})[\w\s]+)\n'"
-	product_replace=r"'Product={{Product}} \1\n'" ## I have added r infront of string to make it raw, otherwise \1 was getting converted to hex
-        role_regex=r"'Role=((?!{{Role}})[\w\s]+)\n'"
-	role_replace=r"'Role={{Role}} \1\n'"
-	line_regex=r"'\[setup\][\n]Product=[\w\s]*[\n]Role=[\w\s]+'"
-	line_line=r"'[setup]\nProduct={{Product}}\nRole={{Role}}'"
-
-	content=content+"""
-  post_tasks:
-
-  - name : Check if {{Dest}} exist's
-    stat: path={{ Dest }}
-    register: st
-
-  - replace : dest={{ Dest }}  regexp=%s  replace=%s
-    when:  st.stat.exists is defined and st.stat.exists==True
-  - replace : dest={{ Dest }}  regexp=%s  replace=%s
-    when:  st.stat.exists is defined and st.stat.exists==True
-
-  - lineinfile : dest={{ Dest }}  regexp=%s  line=%s create=yes mode=644
-    when:  st.stat.exists is defined and st.stat.exists==False
-
-  - setup:
-
-""" %( product_regex,product_replace,role_regex,role_replace,line_regex,line_line)
-
-# repr is used to escape the \n, otherwise it will literally add the newline
-# Complete playbook is written
-
-        fo = open(playbook, "wb")
-        fo.write(content);
-        fo.close()
-
-        cmd="ansible-playbook -i "+ hostfile + " " + playbook
-        if debug:
-                OUT=None
-        else:
-                OUT=PIPE
-        #status,output = commands.getstatusoutput(cmd)
-        #status=subprocess.call(cmd,,stdout=OUT,stderr=OUT,shell=True)
-        p=Popen(cmd.split(),stdout=OUT,stderr=OUT)
-        output=p.communicate()
-        status=p.returncode
-        if status == 0:
-                #Cool playbook run successfull"
-                print ("Successfully Deployed  %s/%s on %s " %(product,role,machine))
-        else:
-                print ("Error:")
-                print (output)
-
-
-        #os.remove(hostfile)
-        #os.remove(playbook)
-        return 0
+        return status
 # +----------------------------------------------------------------------+
 def listroles(args):
 	init()
